rand: bypass jrand48(), broken for GCC 10+ it seems

+++ nimble/drivers/nrf5x/src/ble_phy.c
@@ -1572,11 +1572,11 @@ ble_phy_init(void)
 #ifdef NRF53_SERIES
     /* Errata 158: load trim values after toggling power */
     for (uint32_t index = 0; index < 32ul &&
-         NRF_FICR_NS->TRIMCNF[index].ADDR != 0xFFFFFFFFul; index++) {
-        if (((uint32_t)NRF_FICR_NS->TRIMCNF[index].ADDR & 0xFFFFF000ul) == (volatile uint32_t)NRF_RADIO_NS) {
+          NRF_FICR_NS->TRIMCNF[index].ADDR != 0xFFFFFFFFul; index++) {
+        if ((NRF_FICR_NS->TRIMCNF[index].ADDR & 0xFFFFF000ul) == (volatile uint32_t)NRF_RADIO_NS) {
             *((volatile uint32_t *)NRF_FICR_NS->TRIMCNF[index].ADDR) = NRF_FICR_NS->TRIMCNF[index].DATA;
         }
-    }
+     }
 
     *(volatile uint32_t *)(NRF_RADIO_NS_BASE + 0x774) =
         (*(volatile uint32_t* )(NRF_RADIO_NS_BASE + 0x774) & 0xfffffffe) | 0x01000000;
